misc clean up
