+++ src/classes/GeoPoint.tsx
@@ -1,0 +1,9 @@
+export interface IGeoPoint {
+    latitude: number
+    longitude: number
+}
+
+export class GeoPoint implements IGeoPoint {
+    latitude = 0
+    longitude = 0
+}
